`Refactored routes and components for account and order management`
