+++ src/pages/account/index.tsx
@@ -1,0 +1,5 @@
+const Account = () => {
+  return <div>Account</div>;
+};
+
+export default Account;
